+++ docs/api/geometry.md
@@ -1,0 +1,1 @@
+::: tactics2d.geometry
