+++ docs/api/interpolator.md
@@ -1,0 +1,1 @@
+::: tactics2d.interpolator
