--- docs/api/math.md
@@ -1,9 +1,0 @@
-::: tactics2d.math
-
-::: tactics2d.math.geometry
-    options:
-        heading_level: 2
-
-::: tactics2d.math.interpolate
-    options:
-        heading_level: 2
