+++ mkdocs.yml
@@ -23,8 +23,9 @@ nav:
   - Python API:
     - tactics2d.dataset_parser: "api/dataset_parser.md"
     - tactics2d.envs: "api/envs.md"
+    - tactics2d.geometry: "api/geometry.md"
+    - tactics2d.interpolator: "api/interpolator.md"
     - tactics2d.map: "api/map.md"
-    - tactics2d.math: "api/math.md"
     - tactics2d.participant: "api/participant.md"
     - tactics2d.physics: "api/physics.md"
     - tactics2d.sensor: "api/sensor.md"
