+++ setup.py
@@ -3,7 +3,10 @@ import os
 from pybind11.setup_helpers import Pybind11Extension, build_ext
 from setuptools import setup
 
-include_dirs = ["tactics2d/interpolator/cpp_interpolator/include"]
+include_dirs = [
+    "tactics2d/interpolator/cpp_interpolator/include",
+    "tactics2d/geometry/cpp_geometry/include",
+]
 
 if os.name == "nt":  # Windows
     extra_compile_args = ["/O2", "/D_USE_MATH_DEFINES"]
@@ -18,6 +21,16 @@ ext_modules = [
             "tactics2d/interpolator/cpp_interpolator/src/bezier.cpp",
             "tactics2d/interpolator/cpp_interpolator/src/cubic_spline.cpp",
             "tactics2d/interpolator/cpp_interpolator/src/interpolator_bindings.cpp",
+        ],
+        include_dirs=include_dirs,
+        language="c++",
+        extra_compile_args=extra_compile_args,
+    ),
+    Pybind11Extension(
+        "cpp_geometry",
+        [
+            "tactics2d/geometry/cpp_geometry/src/circle.cpp",
+            "tactics2d/geometry/cpp_geometry/src/geometry_bindings.cpp",
         ],
         include_dirs=include_dirs,
         language="c++",
