+++ tactics2d/geometry/cpp_geometry/src/geometry_bindings.cpp
@@ -1,0 +1,13 @@
+// pybind11_bindings.cpp
+#include <pybind11/pybind11.h>
+
+#include "circle.hpp"
+
+namespace py = pybind11;
+
+PYBIND11_MODULE(cpp_geometry, m) {
+    py::class_<Circle>(m, "Circle")
+        .def_static("get_circle_by_three_points", &Circle::get_circle_by_three_points)
+        .def_static("get_circle_by_tangent_vector",
+                    &Circle::get_circle_by_tangent_vector);
+}
