+++ tactics2d/interpolator/__init__.py
@@ -8,6 +8,8 @@
 from .b_spline import BSpline
 from .bezier import Bezier
 from .cubic_spline import CubicSpline
+from .dubins import Dubins
+from .reeds_shepp import ReedsShepp
 from .spiral import Spiral
 
-__all__ = ["Bezier", "BSpline", "CubicSpline", "Spiral"]
+__all__ = ["BSpline", "Bezier", "CubicSpline", "Dubins", "ReedsShepp", "Spiral"]
