+++ tactics2d/interpolator/bezier.py
@@ -10,41 +10,35 @@ from cpp_interpolator import Bezier as cpp_Bezier
 
 
 class Bezier:
-    """This class implements a Bezier curve interpolator.
+    """This class implements a Bezier curve interpolator."""
 
-    Attributes:
-        order (int): The order of the Bezier curve. The order of a Bezier curve is equal to the number of control points minus one.
-    """
-
-    def __init__(self, order: int):
-        """Initialize the Bezier curve interpolator.
-
-        Args:
-            order (int): The order of the Bezier curve.
-        """
-        self.order = order
-        if self.order < 1:
-            raise ValueError("Bezier interpolator: Order must be greater than or equal to one.")
-
-    def _check_validity(self, control_points: np.ndarray):
-        if len(control_points.shape) != 2 or control_points.shape[1] != 2:
-            raise ValueError("Bezier interpolator: Control points should have shape (n, 2).")
-        if len(control_points) != self.order + 1:
-            raise ValueError(
-                "Bezier interpolator: Number of control points must be equal to order plus one."
-            )
-
-    def get_curve(self, control_points: np.ndarray, n_interpolation: int) -> np.ndarray:
+    @staticmethod
+    def get_curve(
+        control_points: np.ndarray, n_interpolation: int, order: int = None
+    ) -> np.ndarray:
         """
         Args:
             control_points (np.ndarray): The control points of the curve. The shape is (order + 1, 2).
             n_interpolation (int): The number of interpolations.
+            order (int, optional): The order of the Bezier curve. If not provided, it will be set to the number of control points minus one.
 
         Returns:
             curve_points (np.ndarray): The interpolated points of the curve. The shape is (n_interpolation, 2).
         """
-        self._check_validity(control_points)
+        # Check the validity of the inputs
+        if len(control_points.shape) != 2 or control_points.shape[1] != 2:
+            raise ValueError("Control points should have shape (n, 2).")
 
-        # Directly call the optimized C++ function for Bezier curve calculation
-        curve_points = cpp_Bezier.get_curve(control_points.tolist(), n_interpolation)
+        if order is None:
+            order = len(control_points) - 1
+        order = int(order)
+        if order < 1:
+            raise ValueError("Order must be greater than or equal to one.")
+
+        if len(control_points) != order + 1:
+            raise ValueError("Number of control points must be equal to order plus one.")
+
+        # Compute the Bezier curve points using the C++ implementation
+        curve_points = cpp_Bezier.get_curve(control_points.tolist(), int(n_interpolation))
+
         return np.array(curve_points)
