--- tactics2d/math/__init__.py
@@ -1,6 +1,0 @@
-##! python3
-# Copyright (C) 2024, Tactics2D Authors. Released under the GNU GPLv3.
-# @File: __init__.py
-# @Description: Initialize the math module.
-# @Author: Yueyuan Li
-# @Version: 1.0.0
