--- tactics2d/math/cpp_function/CMakeLists.txt
@@ -1,21 +1,0 @@
-cmake_minimum_required(VERSION 3.22.1)
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set(CMAKE_CXX_EXTENSIONS OFF)
-
-include_directories(
-    ${CMAKE_CURRENT_SOURCE_DIR}/include
-    /usr/local/include
-    /usr/include/
-)
-
-
-set(CMAKE_BUILD_TYPE Release)
-set(CMAKE_CXX_FLAGS_RELEASE "-O3")
-
-project(cpp_function)
-
-set(PYBIND11_FINDPYTHON ON)
-find_package(Python REQUIRED COMPONENTS Interpreter Development)
-find_package(pybind11 CONFIG)
-pybind11_add_module(cpp_function circle.cpp b_spline.cpp bezier.cpp cubic_spline.cpp  pybind11_bindings.cpp)
