--- tactics2d/math/geometry/__init__.py
@@ -1,11 +1,0 @@
-##! python3
-# Copyright (C) 2024, Tactics2D Authors. Released under the GNU GPLv3.
-# @File: __init__.py
-# @Description: Initialize the math.geometry module.
-# @Author: Yueyuan Li
-# @Version: 1.0.0
-
-from .circle import Circle
-from .vector import Vector
-
-__all__ = ["Circle", "Vector"]
