--- tactics2d/math/geometry/transform.py
@@ -1,10 +1,0 @@
-def local_to_global():
-    return
-
-
-def global_to_local():
-    return
-
-
-def frenet_projection():
-    return
