--- tactics2d/math/interpolate/__init__.py
@@ -1,15 +1,0 @@
-# #! python3
-# Copyright (C) 2024, Tactics2D Authors. Released under the GNU GPLv3.
-# @File: __init__.py
-# @Description: Initialize the math.interpolate module.
-# @Author: Yueyuan Li
-# @Version: 1.0.0
-
-from .b_spline import BSpline
-from .bezier import Bezier
-from .cubic_spline import CubicSpline
-from .dubins import Dubins
-from .reeds_shepp import ReedsShepp
-from .spiral import Spiral
-
-__all__ = ["BSpline", "Bezier", "CubicSpline", "Dubins", "ReedsShepp", "Spiral"]
